requirements.txt
APScheduler==3.11.0
asgiref==3.8.1
channels==4.2.2
channels_redis==4.2.1
click==8.1.8
colorama==0.4.6
Django==5.2
django-apscheduler==0.7.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
h11==0.16.0
msgpack==1.1.0
pillow==11.2.1
psycopg2-binary==2.9.10
redis==5.2.1
sqlparse==0.5.3
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.34.2
